src/main/webapp/app/entities/hemorragie-visceres/route/hemorragie-visceres-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { ActivatedRoute, ActivatedRouteSnapshot, Router, convertToParamMap } from '@angular/router';
import { of } from 'rxjs';

import { IHemorragieVisceres } from '../hemorragie-visceres.model';
import { HemorragieVisceresService } from '../service/hemorragie-visceres.service';

import hemorragieVisceresResolve from './hemorragie-visceres-routing-resolve.service';

describe('HemorragieVisceres routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let service: HemorragieVisceresService;
  let resultHemorragieVisceres: IHemorragieVisceres | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      providers: [
        provideHttpClient(),
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    service = TestBed.inject(HemorragieVisceresService);
    resultHemorragieVisceres = undefined;
  });

  describe('resolve', () => {
    it('should return IHemorragieVisceres returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        hemorragieVisceresResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultHemorragieVisceres = result;
          },
        });
      });

      // THEN
      expect(service.find).toHaveBeenCalledWith(123);
      expect(resultHemorragieVisceres).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      TestBed.runInInjectionContext(() => {
        hemorragieVisceresResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultHemorragieVisceres = result;
          },
        });
      });

      // THEN
      expect(service.find).not.toHaveBeenCalled();
      expect(resultHemorragieVisceres).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IHemorragieVisceres>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        hemorragieVisceresResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultHemorragieVisceres = result;
          },
        });
      });

      // THEN
      expect(service.find).toHaveBeenCalledWith(123);
      expect(resultHemorragieVisceres).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
